+++ server/db/index.ts
@@ -4,6 +4,7 @@ import * as tenantSchema from './schema/tenant'
 import * as authSchema from './schema/auth'
 import * as unitSchema from './schema/unit'
 import * as accessSchema from './schema/access'
+import * as deviceSchema from './schema/device'
 
 const connectionString = process.env.DATABASE_URL!
 
@@ -15,5 +16,6 @@ export const db = drizzle(client, {
     ...authSchema,
     ...unitSchema,
     ...accessSchema,
+    ...deviceSchema,
   },
 })